requirements.txt
blackjax
fastprogress
ipywidgets
jax
matplotlib
netCDF4
numpyro
pymc >= 5.16.2
scikit-learn
scipy
seaborn
tqdm
xarray
